Moved Comparison and Generation of Speed Dreams Structs from BlackBoxDataTests to separate Classes
