+++ speed-dreams/source-2.2.3/src/simulatedDrivingAssistance/backendUnitTests/CMakeLists.txt
@@ -34,10 +34,16 @@ SET(_SOURCES MsgpackDeserializeTests.cpp
         SQLDatabaseStorageTests.cpp
         BlackBoxDataTests.cpp
         RecorderTests.cpp
+        ComparerUtilsTests.cpp
+            ComparerUtils.cpp
+            GeneratorUtils.cpp
         )
 
 SET(_HEADERS MsgpackSerializeTests.h
-    SocketBlackBoxTests.h)
+    SocketBlackBoxTests.h
+    ComparerUtils.h
+    GeneratorUtils.h
+    )
 
 SET(_MOCK_SOURCES "")
 
@@ -47,7 +53,8 @@ SET(_MOCK_HEADERS
     mocks/SocketBlackBoxMock.h
     mocks/ConfigMock.h
     mocks/PointerManagerMock.h
-    mocks/DecisionMakerMock.h)
+    mocks/DecisionMakerMock.h
+)
 
 
 SD_ADD_EXECUTABLE(backendUnitTests ${_SOURCES} ${_HEADERS} ${_MOCK_SOURCES} ${_MOCK_HEADERS})
