+++ speed-dreams/source-2.2.3/src/simulatedDrivingAssistance/backendUnitTests/ComparerUtilsTests.cpp
@@ -1,0 +1,11 @@
+#include <gtest/gtest.h>
+#include "GeneratorUtils.h"
+#include "ComparerUtils.h"
+
+/// @brief Tests whether CompareSegments will correctly validate it's input.
+TEST(ComparerUtilsTests, SegmentsLongerThanSegmentCount)
+{
+    TestSegments testSegments = Generator::GenerateSegments();
+    EXPECT_THROW(Comparer::CompareSegments(testSegments.NextSegments, testSegments.NextSegments, testSegments.NextSegmentsCount + 1, true), std::invalid_argument);
+    Generator::DestroySegments(testSegments);
+}
